style(std/fields_*): separate Fpk ops specific to choice of pairing
